position/rotation constraint の共通部分を VRM10PostionRotationConstraintEditorBase にまとめた
