--- Assets/VRM10/Runtime/Components/Constraint/IVRM10SourceDestination.cs
@@ -1,11 +1,0 @@
-namespace UniVRM10
-{
-    public interface IVRM10ConstraintSourceDestination
-    {
-        TR GetSourceCoords();
-        TR GetSourceCurrent();
-
-        TR GetDstCoords();
-        TR GetDstCurrent();
-    }
-}
