+++ Assets/VRM10/Runtime/Components/Constraint/VRM10Constraint.cs
@@ -5,9 +5,7 @@ namespace UniVRM10
 {
     public abstract class VRM10Constraint : MonoBehaviour
     {
-        public virtual void Process()
-        {
+        public abstract void Process();
 
-        }
     }
 }
