+++ Assets/VRM10/Runtime/Components/Constraint/VRM10RotationPositionConstraintBase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4d1ea0f0f4d95a2499f9aa5a1d7a2235
+guid: 07919df684c7a4c47a9492a2de417f66
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
